test: split PDF figure candidate resolution coverage
